requirements.txt
blinker==1.9.0
click==8.1.8
colorama==0.4.6
Flask==3.1.0
gunicorn==23.0.0
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
packaging==25.0
Werkzeug==3.1.3
pytesseract==0.3.9
Pillow==9.1.0
pdf2image==1.16.0
python-docx==0.8.11
transformers==4.22.2
torch==1.12.1
python-docx
PyPDF2
textract  # Optional: For OCR fallback
